// next/lib/schema.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<string, bigint | number | string, bigint | number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Accounts {
  access_token: string | null;
  expires_at: Int8 | null;
  id: Generated<number>;
  id_token: string | null;
  provider: string;
  providerAccountId: string;
  refresh_token: string | null;
  scope: string | null;
  session_state: string | null;
  token_type: string | null;
  type: string;
  userId: number;
}

export interface Activations {
  created_at: Generated<Timestamp>;
  created_by: number | null;
  feature_id: string | null;
  id: Generated<string>;
  max_value: number | null;
  max_value_token_index: number | null;
  min_value: number | null;
  sample_id: string | null;
  values: Generated<number[] | null>;
}

export interface Datasets {
  created_at: Generated<Timestamp>;
  created_by: number | null;
  description: string | null;
  id: Generated<string>;
  link: string | null;
  name: string | null;
  type: string | null;
}

export interface Explanations {
  created_at: Generated<Timestamp>;
  created_by: number | null;
  description: string;
  embedding: string | null;
  explanation_model: string | null;
  feature_id: string | null;
  feature_index: number;
  id: Generated<string>;
  layer: number;
  model_id: string;
  notes: string | null;
  type_name: string | null;
  updated_at: Generated<Timestamp>;
}

export interface Features {
  created_at: Generated<Timestamp>;
  created_by: number | null;
  feature_index: number;
  frac_nonzero: Generated<number>;
  id: Generated<string>;
  layer: number;
  model_id: string;
  neg_str: Generated<string[] | null>;
  neg_values: Generated<number[] | null>;
  pos_str: Generated<string[] | null>;
  pos_values: Generated<number[] | null>;
  source_set_name: string | null;
}

export interface Samples {
  created_at: Generated<Timestamp>;
  created_by: number | null;
  dataset_id: string | null;
  id: Generated<string>;
  sample_text: string | null;
  tokens: string[] | null;
}

export interface Sessions {
  expires: Timestamp;
  id: Generated<number>;
  sessionToken: string;
  userId: number;
}

export interface Users {
  email: string | null;
  emailVerified: Timestamp | null;
  id: Generated<number>;
  image: string | null;
  name: string | null;
}

export interface VerificationToken {
  expires: Timestamp;
  identifier: string;
  token: string;
}

export interface DB {
  accounts: Accounts;
  activations: Activations;
  datasets: Datasets;
  explanations: Explanations;
  features: Features;
  samples: Samples;
  sessions: Sessions;
  users: Users;
  verification_token: VerificationToken;
}